Feature/jobs hasan

+++ server/src/modules/jobs/jobs.controller.js
@@ -97,6 +97,9 @@ const updateJob = async (req, res) => {
             return res.status(403).send("Only the owner can edit this Job!");
         }
 
+        if (foundJob.status !== 'open' || foundJob.status !== 'draft') {
+            return res.status(400).json({ err: `Cannot edit this job, due to it being ${foundJob.status}` })
+        }
 
         const updatedJob = await Job.findByIdAndUpdate(req.params.jobId, req.body, { returnDocument: 'after' })
 
@@ -110,6 +113,8 @@ const updateJob = async (req, res) => {
     }
 }
 
+
+
 const deleteJob = async (req, res) => {
     try {
         const foundJob = await Job.findById(req.params.jobId)
@@ -120,6 +125,10 @@ const deleteJob = async (req, res) => {
 
         if (!foundJob.client.equals(req.user._id)) {
             return res.status(403).send("Only the owner can edit this Job!");
+        }
+        
+        if (foundJob.status !== 'draft') {
+            return res.status(400).json({ err: `Cannot edit this job, due to it being ${foundJob.status}` })
         }
 
         const deletedJob = await Job.findByIdAndDelete(req.params.jobId)
@@ -145,7 +154,7 @@ const changeStatus = async (req, res) => {
 
         const updateStatus = { status: req.body.status }
 
-        const updatedJob = await Job.findByIdAndUpdate(req.params.jobId, updateStatus , { returnDocument: 'after' })
+        const updatedJob = await Job.findByIdAndUpdate(req.params.jobId, updateStatus, { returnDocument: 'after' })
 
 
         res.status(200).json(updateJob)
